refactor: extract common code from html doc and frag classes
