+++ lib/loofah/html4/document_fragment.rb
@@ -8,30 +8,7 @@ module Loofah
     #
     class DocumentFragment < Nokogiri::HTML4::DocumentFragment
       include Loofah::TextBehavior
-
-      class << self
-        def parse(tags, encoding = nil)
-          doc = Loofah::HTML4::Document.new
-
-          encoding ||= tags.respond_to?(:encoding) ? tags.encoding.name : "UTF-8"
-          doc.encoding = encoding
-
-          new(doc, tags)
-        end
-      end
-
-      #
-      #  Returns the HTML markup contained by the fragment
-      #
-      def to_s
-        serialize_root.children.to_s
-      end
-
-      alias :serialize :to_s
-
-      def serialize_root
-        at_xpath("./body") || self
-      end
+      include Loofah::HtmlFragmentBehavior
     end
   end
 end
